requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
backports.zoneinfo==0.2.1
black==21.7b0
cachetools==4.2.2
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
importlib-metadata==4.6.4
mypy-extensions==0.4.3
Panda3D==1.10.7
pathspec==0.9.0
python-decouple==3.4
python-telegram-bot==13.7
pytz==2021.1
regex==2021.8.3
six==1.16.0
tomli==1.2.1
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
zipp==3.5.0
